ufm_telemetry_stream_to_fluentd/requirements.txt
requests
configparser
fluent-logger
apscheduler
